app/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic[email]==1.10.14
pymongo==3.13.0
python-dotenv==1.0.1
Jinja2==3.1.2